remove react-window dependency
react-window is used to render the list of cells from the catalogue, but
types definitions are broken with react 18. The number of cells is small
enough that we can remove react-window for the moment. If the number of
cells grows in the future, we can re-implement infinite-scroll or add
pagination.